feat: further reduce complexity of main
